split representation and featurization from `potential` and `utils` in stand alone submodules
