+++ modelforge/tests/test_pairlist.py
@@ -1,0 +1,515 @@
+def test_pairlist_logic():
+    import torch
+
+    # dummy data for illustration
+    positions = torch.tensor(
+        [
+            [0.4933, 0.4460, 0.5762],
+            [0.2340, 0.2053, 0.5025],
+            [0.6566, 0.1263, 0.8792],
+            [0.1656, 0.0338, 0.6708],
+            [0.5696, 0.4790, 0.9622],
+            [0.3499, 0.4241, 0.8818],
+            [0.8400, 0.9389, 0.1888],
+            [0.4983, 0.0793, 0.8639],
+            [0.6605, 0.7567, 0.1938],
+            [0.7725, 0.9758, 0.7063],
+        ]
+    )
+    molecule_indices = torch.tensor(
+        [0, 0, 0, 1, 1, 2, 2, 2, 3, 3]
+    )  # molecule index for each atom
+
+    # generate index grid
+    n = len(molecule_indices)
+    i_indices, j_indices = torch.triu_indices(n, n, 1)
+
+    # filter pairs to only keep those belonging to the same molecule
+    same_molecule_mask = molecule_indices[i_indices] == molecule_indices[j_indices]
+
+    # Apply mask to get final pair indices
+    i_final_pairs = i_indices[same_molecule_mask]
+    j_final_pairs = j_indices[same_molecule_mask]
+
+    # Concatenate to form final (2, n_pairs) tensor
+    final_pair_indices = torch.stack((i_final_pairs, j_final_pairs))
+
+    assert torch.allclose(
+        final_pair_indices,
+        torch.tensor([[0, 0, 1, 3, 5, 5, 6, 8], [1, 2, 2, 4, 6, 7, 7, 9]]),
+    )
+
+    # Create pair_coordinates tensor
+    pair_coordinates = positions[final_pair_indices.T]
+    pair_coordinates = pair_coordinates.view(-1, 2, 3)
+
+    # Calculate distances
+    distances = (pair_coordinates[:, 0, :] - pair_coordinates[:, 1, :]).norm(
+        p=2, dim=-1
+    )
+    # Calculate distances
+    distances = (pair_coordinates[:, 0, :] - pair_coordinates[:, 1, :]).norm(
+        p=2, dim=-1
+    )
+
+    # Define a cutoff
+    cutoff = 1.0
+
+    # Find pairs within the cutoff
+    in_cutoff = (distances <= cutoff).nonzero(as_tuple=False).squeeze()
+
+    # Get the atom indices within the cutoff
+    atom_pairs_withing_cutoff = final_pair_indices[:, in_cutoff]
+    assert torch.allclose(
+        atom_pairs_withing_cutoff,
+        torch.tensor([[0, 0, 1, 3, 5, 5, 8], [1, 2, 2, 4, 6, 7, 9]]),
+    )
+
+
+def test_pairlist():
+    import torch
+
+    from modelforge.potential.models import Neighborlist, Pairlist
+
+    atomic_subsystem_indices = torch.tensor([0, 0, 0, 1, 1, 1])
+    positions = torch.tensor(
+        [
+            [0.0, 0.0, 0.0],
+            [1.0, 1.0, 1.0],
+            [2.0, 2.0, 2.0],
+            [3.0, 3.0, 3.0],
+            [4.0, 4.0, 4.0],
+            [5.0, 5.0, 5.0],
+        ]
+    )
+    from openff.units import unit
+
+    cutoff = unit.Quantity(5.0, unit.nanometer).to(unit.nanometer).m
+    pairlist = Neighborlist(cutoff, only_unique_pairs=True)
+    r = pairlist(positions, atomic_subsystem_indices)
+    pair_indices = r.pair_indices
+
+    # pairlist describes the pairs of interacting atoms within a batch
+    # that means for the pairlist provided below:
+    # pair1: pairlist[0][0] and pairlist[1][0], i.e. (0,1)
+    # pair2: pairlist[0][1] and pairlist[1][1], i.e. (0,2)
+    # pair3: pairlist[0][2] and pairlist[1][2], i.e. (1,2)
+
+    assert torch.allclose(
+        pair_indices, torch.tensor([[0, 0, 1, 3, 3, 4], [1, 2, 2, 4, 5, 5]])
+    )
+    # NOTE: pairs are defined on axis=1 and not axis=0
+    assert torch.allclose(
+        r.r_ij,
+        torch.tensor(
+            [
+                [1.0, 1.0, 1.0],  # pair1, [1.0, 1.0, 1.0] - [0.0, 0.0, 0.0]
+                [2.0, 2.0, 2.0],  # pair2, [2.0, 2.0, 2.0] - [0.0, 0.0, 0.0]
+                [1.0, 1.0, 1.0],  # pair3, [3.0, 3.0, 3.0] - [0.0, 0.0, 0.0]
+                [1.0, 1.0, 1.0],
+                [2.0, 2.0, 2.0],
+                [1.0, 1.0, 1.0],
+            ]
+        ),
+    )
+
+    # test with cutoff
+    cutoff = unit.Quantity(2.0, unit.nanometer).to(unit.nanometer).m
+    pairlist = Neighborlist(cutoff, only_unique_pairs=True)
+    r = pairlist(positions, atomic_subsystem_indices)
+    pair_indices = r.pair_indices
+
+    assert torch.equal(pair_indices, torch.tensor([[0, 1, 3, 4], [1, 2, 4, 5]]))
+    # pairs that are excluded through cutoff: (0,2) and (3,5)
+    assert torch.equal(
+        r.r_ij,
+        torch.tensor(
+            [
+                [1.0, 1.0, 1.0],
+                [1.0, 1.0, 1.0],
+                [1.0, 1.0, 1.0],
+                [1.0, 1.0, 1.0],
+            ]
+        ),
+    )
+
+    assert torch.allclose(
+        r.d_ij, torch.tensor([1.7321, 1.7321, 1.7321, 1.7321]), atol=1e-3
+    )
+
+    # -------------------------------- #
+    # test with complete pairlist
+    cutoff = unit.Quantity(2.0, unit.nanometer).to(unit.nanometer).m
+    neigborlist = Neighborlist(cutoff, only_unique_pairs=False)
+    r = neigborlist(positions, atomic_subsystem_indices)
+    pair_indices = r.pair_indices
+
+    assert torch.equal(
+        pair_indices, torch.tensor([[0, 1, 1, 2, 3, 4, 4, 5], [1, 0, 2, 1, 4, 3, 5, 4]])
+    )
+
+    # -------------------------------- #
+    # make sure that Pairlist and Neighborlist behave the same for large cutoffs
+    cutoff = unit.Quantity(10.0, unit.nanometer).to(unit.nanometer).m
+    only_unique_pairs = False
+    neighborlist = Neighborlist(cutoff, only_unique_pairs=only_unique_pairs)
+    pairlist = Pairlist(only_unique_pairs=only_unique_pairs)
+    r = pairlist(positions, atomic_subsystem_indices)
+    pair_indices = r.pair_indices
+    r = neighborlist(positions, atomic_subsystem_indices)
+    neighbor_indices = r.pair_indices
+
+    assert torch.equal(pair_indices, neighbor_indices)
+
+    # -------------------------------- #
+    # make sure that they are the same also for non-redundant pairs
+    cutoff = unit.Quantity(10.0, unit.nanometer).to(unit.nanometer).m
+    only_unique_pairs = True
+    neighborlist = Neighborlist(cutoff, only_unique_pairs=only_unique_pairs)
+    pairlist = Pairlist(only_unique_pairs=only_unique_pairs)
+    r = pairlist(positions, atomic_subsystem_indices)
+    pair_indices = r.pair_indices
+    r = neighborlist(positions, atomic_subsystem_indices)
+    neighbor_indices = r.pair_indices
+
+    assert torch.equal(pair_indices, neighbor_indices)
+
+    # -------------------------------- #
+    # this should fail
+    cutoff = unit.Quantity(2.0, unit.nanometer).to(unit.nanometer).m
+    only_unique_pairs = True
+    neighborlist = Neighborlist(cutoff, only_unique_pairs=only_unique_pairs)
+    pairlist = Pairlist(only_unique_pairs=only_unique_pairs)
+    r = pairlist(positions, atomic_subsystem_indices)
+    pair_indices = r.pair_indices
+    r = neighborlist(positions, atomic_subsystem_indices)
+    neighbor_indices = r.pair_indices
+
+    assert not pair_indices.shape == neighbor_indices.shape
+
+
+def test_pairlist_precomputation():
+    import numpy as np
+    import torch
+
+    from modelforge.potential.models import Pairlist
+
+    atomic_subsystem_indices = torch.tensor([0, 0, 0])
+
+    pairlist = Pairlist()
+
+    pairs, nr_pairs = pairlist.construct_initial_pairlist_using_numpy(
+        atomic_subsystem_indices.to("cpu")
+    )
+
+    assert pairs.shape == (2, 6)
+    assert nr_pairs[0] == 6
+
+    # 3 molecules, 3 atoms each
+    atomic_subsystem_indices = torch.tensor([0, 0, 0, 1, 1, 1, 2, 2, 2])
+    pairs, nr_pairs = pairlist.construct_initial_pairlist_using_numpy(
+        atomic_subsystem_indices.to("cpu")
+    )
+
+    assert pairs.shape == (2, 18)
+    assert np.all(nr_pairs == [6, 6, 6])
+
+    # 3 molecules, 3,4, and 5 atoms each
+    atomic_subsystem_indices = torch.tensor([0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 2])
+    pairs, nr_pairs = pairlist.construct_initial_pairlist_using_numpy(
+        atomic_subsystem_indices.to("cpu")
+    )
+
+    assert pairs.shape == (2, 38)
+    assert np.all(nr_pairs == [6, 12, 20])
+
+
+def test_pairlist_on_dataset():
+    # Set up a dataset
+    from modelforge.dataset.dataset import DataModule
+    from modelforge.dataset.utils import FirstComeFirstServeSplittingStrategy
+
+    # prepare reference value
+    dataset = DataModule(
+        name="QM9",
+        batch_size=1,
+        version_select="nc_1000_v0",
+        splitting_strategy=FirstComeFirstServeSplittingStrategy(),
+        remove_self_energies=True,
+        regression_ase=False,
+    )
+    dataset.prepare_data()
+    dataset.setup()
+    # -------------------------------#
+    # -------------------------------#
+    # get methane input
+    batch = next(iter(dataset.train_dataloader(shuffle=False))).nnp_input
+    import torch
+
+    # make sure that the pairlist of methane is correct (single molecule)
+    assert torch.equal(
+        batch.pair_list,
+        torch.tensor(
+            [
+                [0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 4, 4, 4, 4],
+                [1, 2, 3, 4, 0, 2, 3, 4, 0, 1, 3, 4, 0, 1, 2, 4, 0, 1, 2, 3],
+            ]
+        ),
+    )
+
+    # test that the pairlist of 2 molecules is correct (which can then be expected also to be true for N molecules)
+    dataset = DataModule(
+        name="QM9",
+        batch_size=2,
+        version_select="nc_1000_v0",
+        splitting_strategy=FirstComeFirstServeSplittingStrategy(),
+        remove_self_energies=True,
+        regression_ase=False,
+    )
+    dataset.prepare_data()
+    dataset.setup()
+    # -------------------------------#
+    # -------------------------------#
+    # get methane input
+    batch = next(iter(dataset.train_dataloader(shuffle=False))).nnp_input
+
+    assert torch.equal(
+        batch.pair_list,
+        torch.tensor(
+            [
+                [
+                    0,
+                    0,
+                    0,
+                    0,
+                    1,
+                    1,
+                    1,
+                    1,
+                    2,
+                    2,
+                    2,
+                    2,
+                    3,
+                    3,
+                    3,
+                    3,
+                    4,
+                    4,
+                    4,
+                    4,
+                    5,
+                    5,
+                    5,
+                    6,
+                    6,
+                    6,
+                    7,
+                    7,
+                    7,
+                    8,
+                    8,
+                    8,
+                ],
+                [
+                    1,
+                    2,
+                    3,
+                    4,
+                    0,
+                    2,
+                    3,
+                    4,
+                    0,
+                    1,
+                    3,
+                    4,
+                    0,
+                    1,
+                    2,
+                    4,
+                    0,
+                    1,
+                    2,
+                    3,
+                    6,
+                    7,
+                    8,
+                    5,
+                    7,
+                    8,
+                    5,
+                    6,
+                    8,
+                    5,
+                    6,
+                    7,
+                ],
+            ]
+        ),
+    )
+
+    # check that the pairlist maximum value for i is the number of atoms in the batch
+    assert (
+        int(batch.pair_list[0][-1].item()) + 1 == 8 + 1 == len(batch.atomic_numbers)
+    )  # +1 because of 0-based indexing
+
+
+def test_inference_neighborlist():
+    from modelforge.potential.models import NeighborlistForInference, Displacement
+    import torch
+
+    coords1 = torch.tensor(
+        [[0, 0, 0], [0, 0, 0], [0.0, 0, 0], [0, 0, 0]], dtype=torch.float32
+    )
+    coords2 = torch.tensor(
+        [[2.0, 0, 0], [-2, 0, 0], [8.0, 0, 0], [-8.0, 0, 0]], dtype=torch.float32
+    )
+
+    coords3 = torch.tensor(
+        [[0.0, 2.0, 0], [0, -2, 0], [0.0, 8.0, 0], [0.0, -8, 0]], dtype=torch.float32
+    )
+
+    coords4 = torch.tensor(
+        [[0.0, 0, 2.0], [0, 0, -2], [0.0, 0, 8.0], [0.0, 0, -8]], dtype=torch.float32
+    )
+
+    box_vectors = torch.tensor(
+        [[10, 0, 0], [0, 10, 0], [0, 0, 10]], dtype=torch.float32
+    )
+
+    displacement_function = Displacement(box_vectors, periodic=True)
+
+    r_ij, d_ij = displacement_function(coords1, coords2)
+
+    assert torch.allclose(
+        r_ij,
+        torch.tensor(
+            [[-2.0, 0, 0], [2.0, 0, 0], [2.0, 0, 0], [-2.0, 0, 0]], dtype=r_ij.dtype
+        ),
+    )
+
+    assert torch.allclose(
+        d_ij, torch.tensor([[2.0], [2.0], [2.0], [2.0]], dtype=d_ij.dtype)
+    )
+
+    r_ij, d_ij = displacement_function(coords1, coords3)
+
+    assert torch.allclose(
+        r_ij,
+        torch.tensor(
+            [[0, -2.0, 0], [0, 2.0, 0], [0, 2.0, 0], [0, -2.0, 0]], dtype=r_ij.dtype
+        ),
+    )
+
+    assert torch.allclose(
+        d_ij, torch.tensor([[2.0], [2.0], [2.0], [2.0]], dtype=d_ij.dtype)
+    )
+
+    r_ij, d_ij = displacement_function(coords1, coords4)
+
+    assert torch.allclose(
+        r_ij,
+        torch.tensor(
+            [[0, 0, -2.0], [0, 0, 2.0], [0, 0, 2.0], [0, 0, -2.0]], dtype=r_ij.dtype
+        ),
+    )
+
+    assert torch.allclose(
+        d_ij, torch.tensor([[2.0], [2.0], [2.0], [2.0]], dtype=d_ij.dtype)
+    )
+
+    coord5 = torch.tensor(
+        [[0.0, 0, 0], [1, 0, 0], [3.0, 0, 0], [8, 0, 0]], dtype=torch.float32
+    )
+
+    nlist = NeighborlistForInference(
+        cutoff=5.0, displacement_function=displacement_function, only_unique_pairs=False
+    )
+
+    from modelforge.dataset.dataset import NNPInput
+
+    data = NNPInput(
+        atomic_numbers=torch.tensor([1, 1, 1, 1], dtype=torch.int64),
+        positions=coord5,
+        atomic_subsystem_indices=torch.tensor([0, 0, 0, 0], dtype=torch.int64),
+        total_charge=torch.tensor([0.0], dtype=torch.float32),
+    )
+
+    pairs, d_ij, r_ij = nlist(data)
+
+    assert pairs.shape[1] == 12
+
+    nlist = NeighborlistForInference(
+        cutoff=5.0, displacement_function=displacement_function, only_unique_pairs=True
+    )
+
+    pairs, d_ij, r_ij = nlist(data)
+
+    assert pairs.shape[1] == 6
+
+    nlist = NeighborlistForInference(
+        cutoff=3.5, displacement_function=displacement_function, only_unique_pairs=False
+    )
+
+    pairs, d_ij, r_ij = nlist(data)
+
+    assert pairs.shape[1] == 10
+
+    assert torch.all(d_ij < 3.5)
+
+
+def test_pairlist_calculate_r_ij_and_d_ij():
+    # Define inputs
+    import torch
+
+    from modelforge.potential.models import Neighborlist
+
+    positions = torch.tensor(
+        [[0.0, 0.0, 0.0], [2.0, 0.0, 0.0], [0.0, 2.0, 0.0], [0.0, 4.0, 1.0]]
+    )
+    atomic_subsystem_indices = torch.tensor([0, 0, 1, 1])
+    from openff.units import unit
+
+    cutoff = unit.Quantity(3.0, unit.nanometer).to(unit.nanometer).m
+
+    # Create Pairlist instance
+    # --------------------------- #
+    # Only unique pairs
+    pairlist = Neighborlist(cutoff, only_unique_pairs=True)
+    pair_indices = pairlist.enumerate_all_pairs(atomic_subsystem_indices)
+
+    # Calculate r_ij and d_ij
+    r_ij = pairlist.calculate_r_ij(pair_indices, positions)
+    d_ij = pairlist.calculate_d_ij(r_ij)
+
+    # Check if the calculated r_ij and d_ij are correct
+    expected_r_ij = torch.tensor([[2.0, 0.0, 0.0], [0.0, 2.0, 1.0]])
+    expected_d_ij = torch.tensor([[2.0000], [2.2361]])
+
+    assert torch.allclose(r_ij, expected_r_ij, atol=1e-3)
+    assert torch.allclose(d_ij, expected_d_ij, atol=1e-3)
+
+    normalized_r_ij = r_ij / d_ij
+    expected_normalized_r_ij = torch.tensor(
+        [[1.0000, 0.0000, 0.0000], [0.0000, 0.8944, 0.4472]]
+    )
+    assert torch.allclose(expected_normalized_r_ij, normalized_r_ij, atol=1e-3)
+
+    # --------------------------- #
+    # ALL pairs
+    pairlist = Neighborlist(cutoff, only_unique_pairs=False)
+    pair_indices = pairlist.enumerate_all_pairs(atomic_subsystem_indices)
+
+    # Calculate r_ij and d_ij
+    r_ij = pairlist.calculate_r_ij(pair_indices, positions)
+    d_ij = pairlist.calculate_d_ij(r_ij)
+
+    # Check if the calculated r_ij and d_ij are correct
+    expected_r_ij = torch.tensor(
+        [[2.0, 0.0, 0.0], [-2.0, 0.0, 0.0], [0.0, 2.0, 1.0], [0.0, -2.0, -1.0]]
+    )
+    expected_d_ij = torch.tensor([[2.0000], [2.0000], [2.2361], [2.2361]])
+
+    assert torch.allclose(r_ij, expected_r_ij, atol=1e-3)
+    assert torch.allclose(d_ij, expected_d_ij, atol=1e-3)
